pyproject.toml
[project]
name = "kyc-intelligence"
version = "1.0.0"
description = "AI-powered KYC client onboarding intelligence system"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "rich>=13.0.0,<14.0.0",
    "fpdf2>=2.7.0,<3.0.0",
    "beautifulsoup4>=4.12.0,<5.0.0",
    "lxml>=5.0.0,<6.0.0",
    "rapidfuzz>=3.0.0,<4.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "PyYAML>=6.0,<7.0",
    "openpyxl>=3.1.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.11.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "types-PyYAML>=6.0.0",
    "pytest-timeout>=2.3.0",
]
encryption = [
    "cryptography>=41.0.0,<44.0.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "generators*", "utilities*", "tools*", "prompts*"]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
warn_return_any = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["utilities.*", "generators.*"]
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["agents.*"]
disallow_untyped_defs = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["anthropic.*", "fpdf.*", "bs4.*", "lxml.*", "rapidfuzz.*", "openpyxl.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]
ignore = ["E501", "B008", "E402", "E741", "SIM102", "SIM108", "SIM115", "SIM117", "UP042"]

[tool.ruff.lint.isort]
known-first-party = ["agents", "utilities", "generators", "tools", "models", "config", "constants"]
